Update oc_rep API
* Store an int64_t in oc_rep_t for integer values. This addresses
  bug LITE-32 which points out that OCF Specs permit values
  in the range [-2^53,2^53] for integers in OCF's data models.
* Add value setter APIs e.g.: oc_rep_set_value_int() for supported
  value types to allow apps to use in conjunction with oc_rep_set_key()
  to encode objects where keys and values are supplied at runtime through
  variables.
* Modified oc_rep_{begin/end}_{object/array}() macros to allow callers to
  directly pass in a pointer to the "parent" parameter. This change also
  makes code look cleaner.
* Updated all sample apps and modules across the stack that were impacted
  by the above changes.

+++ apps/client_multithread_linux.c
@@ -152,7 +152,7 @@ parse_payload(oc_client_response_t *data)
       break;
     case OC_REP_INT:
       printf("%d\n", rep->value.integer);
-      power = rep->value.integer;
+      power = (int)rep->value.integer;
       break;
     case OC_REP_STRING:
       printf("%s\n", oc_string(rep->value.string));
